move genet code out of main cli module
